+++ NEWS.md
@@ -2,6 +2,7 @@
  * Added description for data-import flow
  * Fixed issue with saving ErrorRecords
  * Added batch update of ParsedRecords after assigning Instance id to MARC records
+ * Added support for records processing in XML format
  * Applied parallel approach for mapping from records to instances
  
 ## 2019-05-17 v1.3.1